requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
Pillow
streamlit